% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_shape_extract.R
\name{extract_city_df}
\alias{extract_city_df}
\title{Extract Swiss City Data to JSON}
\usage{
extract_city_df(path, region)
}
\arguments{
\item{path}{String path to the CSV file}

\item{region}{String name of the region to filter}
}
\value{
A JSON string containing city details
}
\description{
Filters Swiss cities by a specific region and converts the selected columns into a JSON format
}
